improve inventory menu
