+++ gradle.properties
@@ -7,7 +7,7 @@ makeevrserg.java.ktarget=17
 # Project
 makeevrserg.project.name=AstraLibs
 makeevrserg.project.group=ru.astrainteractive.astralibs
-makeevrserg.project.version.string=3.0.0-alpha02
+makeevrserg.project.version.string=3.0.0-alpha03
 makeevrserg.project.description=Core utilities for spigot development
 makeevrserg.project.developers=makeevrserg|Makeev Roman|makeevrserg@gmail.com
 makeevrserg.project.url=https://empireprojekt.ru
