+++ menu-bukkit/src/main/kotlin/ru/astrainteractive/astralibs/menu/clicker/MenuClickListener.kt
@@ -1,16 +1,16 @@
 package ru.astrainteractive.astralibs.menu.clicker
 
 import org.bukkit.event.inventory.InventoryClickEvent
-import ru.astrainteractive.astralibs.menu.menu.InventorySlot
-import ru.astrainteractive.astralibs.menu.menu.Menu
+import ru.astrainteractive.astralibs.menu.inventory.InventoryMenu
+import ru.astrainteractive.astralibs.menu.slot.InventorySlot
 
 /**
- * This click listener listens for top inventory which is [Menu] to be clicked
+ * This click listener listens for top inventory which is [InventoryMenu] to be clicked
  */
 class MenuClickListener : ClickListener {
     private val clicksMap = HashMap<Int, Click>()
     override fun onClick(e: InventoryClickEvent) {
-        if (e.clickedInventory?.holder !is Menu) return
+        if (e.clickedInventory?.holder !is InventoryMenu) return
         clicksMap[e.slot]?.invoke(e)
     }
